add limits of int to function exit, modify function ft_atoi

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: adupuy <adupuy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/18 10:16:14 by adupuy            #+#    #+#             */
-/*   Updated: 2020/11/18 10:17:08 by adupuy           ###   ########.fr       */
+/*   Updated: 2021/03/27 10:04:25 by adupuy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int i;
 	int s;
-	int res;
+	long long res;
 
 	i = 0;
 	while ((str[i] >= 9 && str[i] <= 13) || str[i] == 32)
